Bug fix on DefaultBinBuffer.position(int), DefaultBinBuffer.limit(int); Bug fix on BinCodec arrays.

+++ src/main/java/com/jun0rr/jbom/codec/DoubleArrayCodec.java
@@ -46,7 +46,7 @@ public class DoubleArrayCodec implements BinCodec<double[]> {
 
   @Override
   public int calcSize(double[] array) {
-    return Long.BYTES + Double.BYTES + Double.BYTES * array.length;
+    return Long.BYTES + Short.BYTES + Double.BYTES * array.length;
   }
 
 }
